Move enums to FhirUtils
